Improves State.apply performances and adds helper to create prototype
* Helpers

- createPrototypeObject
the generic version to create a Object

- createPrototypeStateAttribute
the version to create prototype for StateAttribute

- createPrototypeNode
the version to create prototype for Node

The reason behind is that we need automatize some id creation for some type,
some extra method on other, so it's easier to have specific method for this.

* Improves State.apply performances

- Adds function to create prototype automatically (createPrototypeClass, StateAttribute, NodeType)
- Uses type id instead of string in StateSet / State to iterate on attribute to apply
- Avoids to push/pop StateSet when using applyStateSet

|----------------+---------------------------+--------|
| tested in dist | withoutpush/pop stateset} | master |
|----------------+---------------------------+--------|
| safari         | 10.44/12                  | 19/20  |
| chrome         | 4.11/4.5                  | 6.7/7  |
| firefox        | 6.8/7.9                   | 12/14  |
|----------------+---------------------------+--------|

Some explanations StateAttribute and filtering type in Compiler

State handle different Stack of StateAttribute and TextureAttribute.
StateAttribute are mostly things like BlendFunc, Depth, ... user can create
custom StateAttribute IF it does not overlap with gl state in another one. eg:
if you create a custom StateAttribute that manipulate BlendFunc, or a
Texture, then the attribute type should be keep in order that the gl state is
not used in different stacks.